[backend] remove cleanMarkings from markingDefinition to avoid circular dependencies
