+++ opencti-platform/opencti-graphql/tests/05-rules/located-at-located-test.js
@@ -99,8 +99,8 @@ describe('Located at located rule', () => {
       expect(parisToEurope.confidence).toBe(38);
       expect(parisToEurope[RULE].length).toBe(2);
       const parisToEuropeMarkings = parisToEurope[RELATION_OBJECT_MARKING];
-      expect(parisToEuropeMarkings.length).toBe(1); // TLP:CLEAR after markings cleaned
-      expect(parisToEuropeMarkings.includes(TLP_CLEAR_INSTANCE.internal_id)).toBeTruthy();
+      expect(parisToEuropeMarkings.length).toBe(1); // TLP:CLEAR
+      expect(parisToEuropeMarkings.includes(TLP_CLEAR_INSTANCE.internal_id)).toBeFalsy();
       expect(parisToEuropeMarkings.includes(TLP_TEST_INSTANCE.internal_id)).toBeTruthy();
       // Remove the relation must remove the inferences
       await internalDeleteElementById(testContext, SYSTEM_USER, parisLocatedToFrance.internal_id);
